refactor: 병원 정보 엔티티 삭제 및 mapper 수정
