--- app/src/main/java/com/example/team25/data/entity/HospitalEntity.kt
@@ -1,13 +1,0 @@
-package com.example.team25.data.entity
-
-import androidx.room.ColumnInfo
-import androidx.room.Entity
-import androidx.room.PrimaryKey
-
-@Entity(tableName = "hospitals")
-data class HospitalEntity(
-    @PrimaryKey(autoGenerate = true) val id: Long = 0,
-    @ColumnInfo(name = "place_id") val placeId: String,
-    @ColumnInfo(name = "name") val name: String,
-    @ColumnInfo(name = "address") val address: String,
-)
