+++ app/src/main/java/com/example/team25/data/entity/mapper/HospitalDomainMapper.kt
@@ -1,0 +1,21 @@
+package com.example.team25.data.entity.mapper
+
+import com.example.team25.data.network.dto.HospitalDto
+import com.example.team25.domain.model.HospitalDomain
+
+object HospitalDomainMapper {
+
+    fun asDomainFromDto(dto: List<HospitalDto>): List<HospitalDomain> {
+        return dto.map { hospitalDto ->
+            HospitalDomain(
+                placeId = hospitalDto.id,
+                name = hospitalDto.name,
+                address = hospitalDto.address,
+            )
+        }
+    }
+}
+
+fun List<HospitalDto>?.asDomainFromDto(): List<HospitalDomain> {
+    return HospitalDomainMapper.asDomainFromDto(this.orEmpty())
+}
